pyproject.toml
[tool.poetry]
name = "ekila-streams"
version = "0.1.0"
description = ""
authors = ["ekila teams"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.2.0"
djangochannelsrestframework = "^1.2.0"
channels-auth-token-middlewares = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest-django = "^4.5.2"
pytest = "^7.4.2"
factory-boy = "^3.3.0"
black = "^23.9.1"
django-silk = "^5.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ekila_streams.settings"
